What challenges did I face?
Shortcuts such as @/ do not work in Node when running a script outside of the Next.js environment — I had to replace them with relative paths manually.

The seed.ts script was generating new IDs that did not match the IDs in mockListings.ts. The solution? I removed the IDs from the mock data and had Prisma generate them, then another script extracts them and updates the file automatically.

Runtime tools such as ts-node did not work with --loader, so I learned to use tsx, which is lighter and easier to work with TypeScript📌🤗️